Moved Coord to the new coord.h file.
