+++ graphics/graphics.h
@@ -1,34 +1,9 @@
-#include <cairo.h>
 #include <memory>
 #include <list>
 #include <sstream>
+#include "coord.h"
 
 using namespace std;
-
-class Coord
-{
-public:
-    Coord (double x, double y): _x(x), _y(y) {}
-
-    virtual double x() const { return _x; }
-    virtual double y() const { return _y; }
-
-    // Scale x by factor fx, y by factor fy.
-    Coord scale(double fx, double fy)
-    {
-        return Coord(x() * fx, y() * fy);
-    }
-
-    // Move by dx horizontally, dy vertically.
-    Coord translate(double dx, double dy)
-    {
-        return Coord(x() + dx, y() + dy);
-    }
-
-private:
-    double _x;
-    double _y;
-};
 
 // Drawable area of the screen
 class Canvas
